redis auto-变量命名

+++ carrot/src/main/java/com/example/carrot/autoconfig/redis/RedisAutoConfiguration.java
@@ -32,7 +32,7 @@ public class RedisAutoConfiguration {
         JedisPoolConfig config = new JedisPoolConfig();
         config.setMaxTotal(this.properties.getMaxTotal());
         config.setMaxIdle(this.properties.getMaxIdle());
-        config.setMaxWaitMillis((long)this.properties.getMaxWaitMillis());
+        config.setMaxWaitMillis(this.properties.getMaxWaitMillis());
         config.setTestOnBorrow(this.properties.isTestOnBorrow());
         return config;
     }
